pyproject.toml
[project]
name = "direct_webapp"
dynamic = ["version"]
description = "A Django webapp for hosting the Digital Research Competencies Framework"
authors = [
    { name = "Adrian D'Alessandro", email = "a.dalessandro@imperial.ac.uk" },
]
requires-python = ">=3.12"
dependencies = [
    "django",
    "django-bootstrap5",
    "django-stubs-ext",
    "whitenoise",
    "sendgrid",
    "pillow",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pip-tools",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "pytest-mock",
    "django-stubs[compatible-mypy]",
    "djlint",
]
doc = [
    "mkdocs",
    "mkdocstrings",
    "mkdocstrings-python",
    "mkdocs-material",
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
    "mkdocs-section-index",
]

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = [
    "htmlcov",
] # Exclude the coverage report file from setuptools package finder

[tool.mypy]
disallow_any_explicit = true
disallow_any_generics = true
warn_unreachable = true
warn_unused_ignores = true
disallow_untyped_defs = true
exclude = [".venv/", "docs/", "manage.py", "main/migrations/"]
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
addopts = "-v -p no:warnings --cov=. --cov-report=html --doctest-modules --ignore=direct_webapp/__main__.py --ignore=docs --ignore=direct_webapp/settings/"
DJANGO_SETTINGS_MODULE = "direct_webapp.settings"
FAIL_INVALID_TEMPLATE_VARS = true

[tool.ruff]
exclude = ["main/migrations"]
target-version = "py312"

[tool.ruff.lint]
select = [
    "D",   # pydocstyle
    "E",   # pycodestyle
    "F",   # Pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "RUF", # ruff
]
pydocstyle.convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D104"] # Missing docstring in public package

[tool.django-stubs]
django_settings_module = "direct_webapp.settings"

[tool.coverage.run]
omit = ["tests/*"]

[tool.djlint]
use_gitignore = true
# ignore inline styles
# ignore missing image width and height
ignore="H021,H006"

[tool.djlint.per-file-ignores]
"main/templates/main/base.html"= "H030, H031, H016" # meta tags and title is in head.html
